move to llm utils
