+++ llama-index-integrations/agent/llama-index-agent-openai/llama_index/agent/openai/utils.py
@@ -1,5 +1,5 @@
 """Utils for OpenAI agent."""
 
-from llama_index.utils.openai import resolve_tool_choice
+from llama_index.llms.openai.utils import resolve_tool_choice
 
 __all__ = ["resolve_tool_choice"]
