+++ llama-index-integrations/llms/llama-index-llms-openai/llama_index/llms/openai/utils.py
@@ -428,3 +428,14 @@ def validate_openai_api_key(api_key: Optional[str] = None) -> None:
 
     if not openai_api_key:
         raise ValueError(MISSING_API_KEY_ERROR_MESSAGE)
+
+
+def resolve_tool_choice(tool_choice: Union[str, dict] = "auto") -> Union[str, dict]:
+    """Resolve tool choice.
+
+    If tool_choice is a function name string, return the appropriate dict.
+    """
+    if isinstance(tool_choice, str) and tool_choice not in ["none", "auto", "required"]:
+        return {"type": "function", "function": {"name": tool_choice}}
+
+    return tool_choice
