+++ llama-index-integrations/llms/llama-index-llms-openai/pyproject.toml
@@ -29,13 +29,12 @@ exclude = ["**/BUILD"]
 license = "MIT"
 name = "llama-index-llms-openai"
 readme = "README.md"
-version = "0.2.8"
+version = "0.2.9"
 
 [tool.poetry.dependencies]
 python = ">=3.8.1,<4.0"
 openai = "^1.40.0"
 llama-index-core = "^0.11.7"
-llama-index-utils-openai = "^0.1.0"
 
 [tool.poetry.group.dev.dependencies]
 ipython = "8.10.0"
