--- llama-index-utils/llama-index-utils-openai/BUILD
@@ -1,3 +1,0 @@
-poetry_requirements(
-    name="poetry",
-)
