--- llama-index-utils/llama-index-utils-openai/llama_index/utils/openai/BUILD
@@ -1,1 +1,0 @@
-python_sources()
